vue notifications and some cleanup. a lot of the work done for responsiveness. contact us modal

+++ resources/js/components/blog/BlogShowComment.vue
@@ -27,26 +27,26 @@
     <div class="col-md-9">
         <div class="comment-section p-2">
             <div class="row">
-                <div class="col-md-4">
+                <div class="col-md-4 mt-md-0 mt-3">
                     <label class="form-label" for="title">Display name *</label>
                     <input class="form-control form-control-custom" :class="{ 'is-invalid': errors.name }" id="title"
-                           v-model="commentCreateForm.name">
+                           v-model="commentCreateForm.name" placeholder="Enter display name">
                     <div v-if="errors.name" class="invalid-feedback">
                         {{ errors.name[0] }}
                     </div>
                 </div>
-                <div class="col-md-4">
+                <div class="col-md-4 mt-md-0 mt-3">
                     <label class="form-label" for="email">Email address *</label>
                     <input class="form-control form-control-custom" :class="{ 'is-invalid': errors.email }" id="email"
-                           v-model="commentCreateForm.email">
+                           v-model="commentCreateForm.email" placeholder="Enter email">
                     <div v-if="errors.email" class="invalid-feedback">
                         {{ errors.email[0] }}
                     </div>
                 </div>
-                <div class="col-md-4">
+                <div class="col-md-4 mt-md-0 mt-3">
                     <label class="form-label" for="website">Website</label>
                     <input class="form-control form-control-custom" :class="{ 'is-invalid': errors.website }"
-                           id="website" v-model="commentCreateForm.website">
+                           id="website" v-model="commentCreateForm.website" placeholder="www.yourwebsite.com">
                     <div v-if="errors.website" class="invalid-feedback">
                         {{ errors.website[0] }}
                     </div>
@@ -56,16 +56,15 @@
                 <div class="col-md-9">
                     <textarea class="form-control form-control-custom" :class="{ 'is-invalid': errors.comment }"
                               id="comment" rows="3"
-                              v-model="commentCreateForm.comment"></textarea>
+                              v-model="commentCreateForm.comment" placeholder="Enter your comment here :)"></textarea>
                     <div v-if="errors.comment" class="invalid-feedback">
                         {{ errors.comment[0] }}
                     </div>
                 </div>
-                <div class="col-md-3">
+                <div class="col-md-3 mt-md-0 mt-3">
                     <button type="submit" class="btn btn-custom-neutral w-100 h-100" @click="postComment">Comment
                     </button>
                 </div>
-                <div>is admin: {{ this.isAdmin }}</div>
             </div>
         </div>
     </div>
@@ -79,8 +78,9 @@ import {useAuthStore} from "../../store/modules/Auth";
 import {createPinia} from 'pinia';
 
 const pinia = createPinia()
-
 const authStore = useAuthStore(pinia);
+
+import Notification from "./../Helpers/Notification";
 
 export default {
     name: "BlogShowComment",
@@ -116,10 +116,13 @@ export default {
             });
         },
         deleteComment(comment) {
-            BlogService.deleteComment(this.blog, comment).then(res => {
-                this.getComments();
-            }).catch(e => {
-                this.errors = e.response.data.errors;
+            Notification.methods.confirm().then(() => {
+                BlogService.deleteComment(this.blog, comment).then(res => {
+                    this.getComments();
+                    Notification.methods.success('Comment deleted');
+                }).catch(e => {
+                    this.errors = e.response.data.errors;
+                });
             });
         },
         resetFormFields() {
@@ -148,9 +151,9 @@ export default {
 </script>
 
 <style scoped>
-    .remove-comment {
-        text-decoration: none;
-        color: #FF5161;
-        cursor: pointer;
-    }
+.remove-comment {
+    text-decoration: none;
+    color: #FF5161;
+    cursor: pointer;
+}
 </style>
